Updates per code review for #4
