+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-01-12
+
+### Added
+- Support to disable 2FA for API token generation ([#1](https://github.com/markshust/magento2-module-disabletwofactorauth/issues/1)).
+
+### Updated
+- Updated docblocks and other minor formatting issues.
+- Updated REAMDE to make it more explicit not to disable 2FA within production environments.
+
 ## [1.0.0] - 2020-08-10
 
 ### Added
