+++ Plugin/BypassTwoFactorAuth.php
@@ -6,11 +6,21 @@ namespace MarkShust\DisableTwoFactorAuth\Plugin;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\TwoFactorAuth\Model\TfaSession;
 
+/**
+ * Class BypassTwoFactorAuth
+ * @package MarkShust\DisableTwoFactorAuth\Plugin
+ */
 class BypassTwoFactorAuth
 {
-    /** @var ScopeConfigInterface */
-    private $scopeConfig;
+    const XML_PATH_CONFIG_ENABLE = 'twofactorauth/general/enable';
 
+    /** @var ScopeConfigInterface */
+    private ScopeConfigInterface $scopeConfig;
+
+    /**
+     * BypassTwoFactorAuth constructor.
+     * @param ScopeConfigInterface $scopeConfig
+     */
     public function __construct(
         ScopeConfigInterface $scopeConfig
     ) {
@@ -18,17 +28,24 @@ class BypassTwoFactorAuth
     }
 
     /**
-     * If the TwoFactorAuth module Enable setting is set to false, always return true here so all requests bypass 2FA.
-     * Otherwise, return the original result.
+     * Enables the bypass of 2FA for admin access.
+     * This can be useful within development & integration environments.
+     *
+     * If 2FA is enabled, return the original result.
+     * If 2FA is disabled, always return true so all requests bypass 2FA.
+     *
+     * NOTE: Always keep 2FA enabled within production environments for security purposes.
      *
      * @param TfaSession $subject
      * @param $result
      * @return bool
      */
-    public function afterIsGranted(TfaSession $subject, $result): bool
-    {
-        return !$this->scopeConfig->isSetFlag('twofactorauth/general/enable')
-            ? true
-            : $result;
+    public function afterIsGranted(
+        TfaSession $subject,
+        $result
+    ): bool {
+        return $this->scopeConfig->isSetFlag(self::XML_PATH_CONFIG_ENABLE)
+            ? $result
+            : true;
     }
 }
